scripts/03_hdwgcna/hdwgcna_analysis.py
"""
HDWGCNA (Hierarchical Dynamics Weighted Gene Co-expression Network Analysis)
共表达网络分析脚本
"""

import argparse
import os
import yaml
from pathlib import Path
from loguru import logger
import numpy as np
import pandas as pd


def setup_logging(log_dir):
    """设置日志"""
    os.makedirs(log_dir, exist_ok=True)
    logger.add(
        os.path.join(log_dir, "hdwgcna_{time}.log"),
        format="{time:YYYY-MM-DD HH:mm:ss} | {level: <8} | {message}",
        level="INFO"
    )


def load_expression_matrix(input_file):
    """
    加载表达矩阵
    
    Parameters
    ----------
    input_file : str
        表达矩阵文件路径
        
    Returns
    -------
    pd.DataFrame
        基因 × 样本的表达矩阵
    """
    logger.info(f"加载表达矩阵: {input_file}")
    
    if input_file.endswith('.csv'):
        expr_matrix = pd.read_csv(input_file, index_col=0)
    elif input_file.endswith('.xlsx'):
        expr_matrix = pd.read_excel(input_file, index_col=0)
    elif input_file.endswith('.h5ad'):
        import scanpy as sc
        adata = sc.read_h5ad(input_file)
        expr_matrix = pd.DataFrame(
            adata.X.toarray() if hasattr(adata.X, 'toarray') else adata.X,
            index=adata.var_names,
            columns=adata.obs_names
        )
    else:
        raise ValueError(f"不支持的文件格式: {input_file}")
    
    logger.info(f"表达矩阵形状: {expr_matrix.shape}")
    
    return expr_matrix


def preprocess_expression(expr_matrix, config):
    """
    表达矩阵预处理
    
    Parameters
    ----------
    expr_matrix : pd.DataFrame
        表达矩阵
    config : dict
        配置参数
        
    Returns
    -------
    pd.DataFrame
        预处理后的表达矩阵
    """
    logger.info("进行表达矩阵预处理")
    
    # 移除低表达基因
    expr_matrix = expr_matrix[(expr_matrix > 0).sum(axis=1) >= 10]
    logger.info(f"过滤后基因数: {expr_matrix.shape[0]}")
    
    # 标准化
    expr_matrix = (expr_matrix - expr_matrix.mean()) / expr_matrix.std()
    
    return expr_matrix


def network_construction(expr_matrix, config):
    """
    构建基因共表达网络
    
    Parameters
    ----------
    expr_matrix : pd.DataFrame
        表达矩阵
    config : dict
        配置参数
        
    Returns
    -------
    pd.DataFrame
        相似度矩阵
    """
    logger.info("构建基因共表达网络")
    
    # TODO: 实现WGCNA网络构建逻辑
    # 这部分通常在R中实现，可以调用R接口或在Python中重新实现
    
    wgcna_config = config['wgcna']
    logger.info(f"软功率: {wgcna_config['soft_power']}")
    logger.info(f"最小模块大小: {wgcna_config['min_module_size']}")
    
    return None


def module_detection(expr_matrix, config):
    """
    检测共表达模块
    
    Parameters
    ----------
    expr_matrix : pd.DataFrame
        表达矩阵
    config : dict
        配置参数
        
    Returns
    -------
    pd.DataFrame
        基因模块分配
    """
    logger.info("检测共表达模块")
    
    # TODO: 实现模块检测逻辑
    
    return None


def module_annotation(modules, expr_matrix, config):
    """
    模块功能注释
    
    Parameters
    ----------
    modules : pd.DataFrame
        基因模块分配
    expr_matrix : pd.DataFrame
        表达矩阵
    config : dict
        配置参数
        
    Returns
    -------
    pd.DataFrame
        模块注释结果
    """
    logger.info("进行模块功能注释")
    
    # TODO: 实现模块注释逻辑
    
    return None


def main():
    parser = argparse.ArgumentParser(description="HDWGCNA共表达分析")
    parser.add_argument(
        "--input",
        required=True,
        help="输入表达矩阵文件"
    )
    parser.add_argument(
        "--config",
        required=True,
        help="配置文件路径"
    )
    parser.add_argument(
        "--output",
        default="../results/",
        help="输出目录"
    )
    
    args = parser.parse_args()
    
    # 加载配置
    with open(args.config, 'r') as f:
        config = yaml.safe_load(f)
    
    # 设置日志
    setup_logging(config['paths']['logs_dir'])
    
    logger.info("=" * 50)
    logger.info("HDWGCNA共表达分析启动")
    logger.info("=" * 50)
    
    # 加载数据
    expr_matrix = load_expression_matrix(args.input)
    
    # 预处理
    expr_matrix = preprocess_expression(expr_matrix, config)
    
    # 网络构建
    network = network_construction(expr_matrix, config)
    
    # 模块检测
    modules = module_detection(expr_matrix, config)
    
    # 模块注释
    module_anno = module_annotation(modules, expr_matrix, config)
    
    logger.info("分析完成！")


if __name__ == "__main__":
    main()
