requirements.txt
# 单细胞分析核心库
scanpy>=1.9.0
anndata>=0.8.0
scipy>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 数据处理和可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
adjustText>=0.7.3

# 生物信息学工具
biopython>=1.79
pysam>=0.16.0

# 统计分析
statsmodels>=0.13.0
scikit-posthocs>=0.4.0

# 并行处理
joblib>=1.1.0

# 文件处理
pyyaml>=5.4
openpyxl>=3.7.0

# 日志和进度条
tqdm>=4.62.0
loguru>=0.5.3

# Jupyter支持（可选）
jupyter>=1.0.0
notebook>=6.4.0
ipython>=7.30.0
